requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
Django==5.1.7
dotenv==0.9.9
h11==0.14.0
idna==3.10
mysql-connector-python==9.2.0
outcome==1.3.0.post0
packaging==24.2
pycparser==2.22
PyMySQL==1.1.1
PySocks==1.7.1
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
selenium==4.29.0
sniffio==1.3.1
sortedcontainers==2.4.0
spotipy==2.25.1
sqlparse==0.5.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
